Apply suggested changes to added test
